requirements.txt
Django>=4.0
djangorestframework>=3.14
python-decouple>=3.8
requests>=2.28
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=4.5
